Enum memoization::Memoized

pub enum Memoized<I: 'static, O, Func: Fn(I) -> O> {
    UnInitialized(PhantomData<&'static I>, Box<Func>),
    Processed(O),
}

Magical Enum

The generic enum that allows for memoization to happeni.

  use memoization::Memoized;

  struct Example<I:'static, O: Clone, F: Fn(I) -> O> {
       data: Memoized<I,O,F>
  }

  fn eq_str(a: &str, b: &str) -> bool {
        a == b
  }

  let lambda = |x: i32 | -> String {
        x.to_string()
  };
  let mut dut = Example {
        data: Memoized::new(lambda)
  };
  //field is memoized but it can still be written too.
  *dut.data = "9001".to_string();
  //field can be borrowed as its output data type.
  assert!( eq_str( &dut.data, "9001") );

Variants
UnInitialized(PhantomData<&'static I>, Box<Func>)Processed(O)
Methods

impl<I: 'static, O, Func: Fn(I) -> O> Memoized<I, O, Func>

fn new(lambda: Func) -> Memoized<I, O, Func>

Build a new memoized field. The user will pass a lambda function that will initialize the field.

 use memoization::Memoized;

 let lambda = | a: (i32,i64,&str) | -> String {
       format!("Line {:?} DateCode {:?} Log \"{}\"",a.0,a.1,a.2)
 };

 let memoized = Memoized::new(lambda);

fn process(&mut self, data: I)

This will convert an UnInitialized value to a Processed value. When called on a Processed value this function will PANIC.

 use memoization::Memoized;

 let tup: (i32,i64,&'static str) = (20,-15,"Hello World!");
 let lambda = | a: (i32,i64,&str) | -> String {
       format!("Line {:?} DateCode {:?} Log \"{}\"",a.0,a.1,a.2)
 };
 let mut memoized = Memoized::new(lambda);
 //process the data
 memoized.process(tup);
 //borrowing the processed, as it's processed data type
 let x: &str = &memoized;
 assert_eq!( x, "Line 20 DateCode -15 Log \"Hello World!\"");

fn processed(&self) -> bool

Informs user if a field has been Processed.

 use memoization::Memoized;

 let tup: (i32,i64,&'static str) = (20,-15,"Hello World!");
 let lambda = | a: (i32,i64,&str) | -> String {
       format!("Line {:?} DateCode {:?} Log \"{}\"",a.0,a.1,a.2)
 };
 let mut memoized = Memoized::new(lambda);
 //data is not initalized/processed
 assert!( ! memoized.processed() );
 //process the data
 memoized.process(tup);
 //data is now initalized
 assert!( memoized.processed() );